requirements.txt
pinecone-client 
openai 
tiktoken 
langchain
pyinstrument

